@@ -68,7 +68,7 @@
     "react-translate-component": "^0.13.2",
     "svgo": "^0.5.1",
     "tcomb": "2.5.2",
-    "whatwg-fetch": "^0.11.0"
+    "whatwg-fetch": "^2.0.1"
   },
   "devDependencies": {
     "babel-core": "^6.21.0",
@@ -89,25 +89,23 @@
     "eslint-plugin-react": "^5.2.2",
     "express": "^4.14.0",
     "express-history-api-fallback": "^2.1.0",
-    "extract-text-webpack-plugin": "^0.9.1",
+    "extract-text-webpack-plugin": "^2.0.0-beta.4",
     "file-loader": "^0.8.4",
     "jest-cli": "~0.4.1",
-    "json-loader": "^0.5.3",
+    "json-loader": "^0.5.4",
     "mocha": "^2.5.3",
     "node-libs-browser": "^1.0.0",
     "node-sass": "^3.4.1",
     "postcss-loader": "^0.8.2",
-    "react-transform-hmr": "^1.0.4",
-    "remarkable-loader": "^0.1.0",
-    "sass-loader": "^3.1.1",
+    "remarkable-loader": "^0.2.1",
+    "sass-loader": "^4.1.1",
     "script-loader": "^0.6.1",
-    "style-loader": "^0.12.4",
+    "style-loader": "^0.13.1",
     "svg-inline-loader": "^0.3.0",
     "svgo-loader": "^1.1.0",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
+    "webpack": "^2.2.0-rc.3",
     "webpack-dev-middleware": "^1.9.0",
-    "webpack-dev-server": "^1.14.1",
     "webpack-hot-middleware": "^2.14.0",
     "worker-loader": "^0.6.0"
   },
