@@ -59,4 +59,4 @@ class BlockchainActions {
 const BlockchainActionsInstance = alt.createActions(BlockchainActions);
 Apis.setRpcConnectionStatusCallback(BlockchainActionsInstance.updateRpcConnectionStatus);
 
-module.exports = BlockchainActionsInstance;
+export default BlockchainActionsInstance;
