@@ -1,10 +1,11 @@
+if (module.hot) module.hot.accept();
 require("./assets/loader-dev");
 if (!window.Intl) { // Safari polyfill
     require.ensure(['intl'], require => {
     	window.Intl = require('intl');
-        Intl.__addLocaleData(require("./assets/intl-data/en.json"));        
-        require("App.jsx");
+        Intl.__addLocaleData(require("./assets/intl-data/en.json"));
+        require("index.js");
     });
 } else {
-    require("App.jsx");
+    require("index.js");
 }
