@@ -1,28 +1,28 @@
 // Core asset
-require("file?name=asset-symbols/[name].png!./bts.png");
+require("file-loader?name=asset-symbols/[name].png!./bts.png");
 
 // BitAssets
-require("file?name=asset-symbols/[name].png!./usd.png");
-require("file?name=asset-symbols/[name].png!./eur.png");
-require("file?name=asset-symbols/[name].png!./cny.png");
-require("file?name=asset-symbols/[name].png!./gold.png");
-require("file?name=asset-symbols/[name].png!./btc.png");
-require("file?name=asset-symbols/silver.png!./bts.png");
+require("file-loader?name=asset-symbols/[name].png!./usd.png");
+require("file-loader?name=asset-symbols/[name].png!./eur.png");
+require("file-loader?name=asset-symbols/[name].png!./cny.png");
+require("file-loader?name=asset-symbols/[name].png!./gold.png");
+require("file-loader?name=asset-symbols/[name].png!./btc.png");
+require("file-loader?name=asset-symbols/silver.png!./bts.png");
 
 // 3rd party assets
-require("file?name=asset-symbols/[name].png!./eth.png");
-require("file?name=asset-symbols/[name].png!./steem.png");
-require("file?name=asset-symbols/[name].png!./mkr.png");
-require("file?name=asset-symbols/[name].png!./dgd.png");
-require("file?name=asset-symbols/[name].png!./obits.png");
-require("file?name=asset-symbols/[name].png!./btsr.png");
-require("file?name=asset-symbols/[name].png!./dao.png");
-require("file?name=asset-symbols/[name].png!./lisk.png");
-require("file?name=asset-symbols/[name].png!./peerplays.png");
-require("file?name=asset-symbols/[name].png!./icoo.png");
-require("file?name=asset-symbols/[name].png!./blockpay.png");
-require("file?name=asset-symbols/[name].png!./dash.png");
-require("file?name=asset-symbols/[name].png!./eurt.png");
-require("file?name=asset-symbols/[name].png!./game.png");
-require("file?name=asset-symbols/[name].png!./grc.png");
-require("file?name=asset-symbols/[name].png!./usdt.png");
+require("file-loader?name=asset-symbols/[name].png!./eth.png");
+require("file-loader?name=asset-symbols/[name].png!./steem.png");
+require("file-loader?name=asset-symbols/[name].png!./mkr.png");
+require("file-loader?name=asset-symbols/[name].png!./dgd.png");
+require("file-loader?name=asset-symbols/[name].png!./obits.png");
+require("file-loader?name=asset-symbols/[name].png!./btsr.png");
+require("file-loader?name=asset-symbols/[name].png!./dao.png");
+require("file-loader?name=asset-symbols/[name].png!./lisk.png");
+require("file-loader?name=asset-symbols/[name].png!./peerplays.png");
+require("file-loader?name=asset-symbols/[name].png!./icoo.png");
+require("file-loader?name=asset-symbols/[name].png!./blockpay.png");
+require("file-loader?name=asset-symbols/[name].png!./dash.png");
+require("file-loader?name=asset-symbols/[name].png!./eurt.png");
+require("file-loader?name=asset-symbols/[name].png!./game.png");
+require("file-loader?name=asset-symbols/[name].png!./grc.png");
+require("file-loader?name=asset-symbols/[name].png!./usdt.png");
