@@ -86,14 +86,14 @@ let CallOrder = t.struct({
 }, "CallOrder");
 
 
-module.exports = {
-    Asset: Asset,
-    Block: Block,
-    WalletTcomb: WalletTcomb,
+export {
+    Asset,
+    Block,
+    WalletTcomb,
     //PublicKeyTcomb: PublicKeyTcomb,
-    PrivateKeyTcomb: PrivateKeyTcomb,
-    LimitOrder: LimitOrder,
-    ShortOrder: ShortOrder,
-    CallOrder: CallOrder,
-    SettleOrder: SettleOrder
+    PrivateKeyTcomb,
+    LimitOrder,
+    ShortOrder,
+    CallOrder,
+    SettleOrder
 };
